Move AHC sklearn code to its own folder
